# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReadinessStatusChecks = {
  postgres: string;
  redis: string;
};

export interface ReadinessStatus {
  status: string;
  checks: ReadinessStatusChecks;
}

/**
 * A workflow and its metadata (not its graph — see WorkflowVersion).
 */
export interface Workflow {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  tags: string[];
  isActive: boolean;
  /** @nullable */
  activeVersionId: string | null;
  /**
     * When this workflow's most recent execution was created. Null if it has never run.
     * @nullable
     */
  lastExecutionAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type WorkflowGraphNodeType = typeof WorkflowGraphNodeType[keyof typeof WorkflowGraphNodeType];


export const WorkflowGraphNodeType = {
  start: 'start',
  webhook_trigger: 'webhook_trigger',
  http_request: 'http_request',
  delay: 'delay',
  if: 'if',
  end: 'end',
} as const;

export type WorkflowGraphNodePosition = {
  x: number;
  y: number;
};

/**
 * Node-type-specific settings (shape depends on type).
 */
export type WorkflowGraphNodeConfig = { [key: string]: unknown };

/**
 * A single node placed on the canvas.
 */
export interface WorkflowGraphNode {
  /** Stable node identifier, unique within the graph. */
  key: string;
  type: WorkflowGraphNodeType;
  /** @nullable */
  label?: string | null;
  position?: WorkflowGraphNodePosition;
  /** Node-type-specific settings (shape depends on type). */
  config?: WorkflowGraphNodeConfig;
}

/**
 * A directed edge between two nodes.
 */
export interface WorkflowGraphConnection {
  sourceKey: string;
  /** @nullable */
  sourceHandle?: string | null;
  targetKey: string;
  /** @nullable */
  targetHandle?: string | null;
}

export interface WorkflowGraph {
  nodes: WorkflowGraphNode[];
  connections: WorkflowGraphConnection[];
}

export interface WorkflowInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  description?: string;
  tags?: string[];
  graph?: WorkflowGraph;
}

/**
 * At least one field must be provided.
 */
export interface WorkflowUpdate {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name?: string;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  isActive?: boolean;
}

export interface WorkflowVersionInput {
  graph: WorkflowGraph;
  description?: string;
}

/**
 * A full, immutable snapshot of a workflow's graph.
 */
export interface WorkflowVersion {
  id: string;
  workflowId: string;
  version: number;
  graphJson: WorkflowGraph;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

/**
 * Version metadata without the (potentially large) graph payload.
 */
export interface WorkflowVersionSummary {
  id: string;
  version: number;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

export type ExecutionStatus = typeof ExecutionStatus[keyof typeof ExecutionStatus];


export const ExecutionStatus = {
  pending: 'pending',
  running: 'running',
  success: 'success',
  error: 'error',
  cancelled: 'cancelled',
  timeout: 'timeout',
} as const;

export type ExecutionTriggerType = typeof ExecutionTriggerType[keyof typeof ExecutionTriggerType];


export const ExecutionTriggerType = {
  manual: 'manual',
  webhook: 'webhook',
  schedule: 'schedule',
  api: 'api',
} as const;

/**
 * A single run of a workflow.
 */
export interface Execution {
  id: string;
  workflowId: string;
  versionId: string;
  status: ExecutionStatus;
  triggerType: ExecutionTriggerType;
  /** The input the entry node ran with. Arbitrary JSON, or null. */
  triggerPayload: unknown;
  /** The workflow's final output (its terminal node's output, or one merged object keyed by node if it has several). Null until the execution finishes. */
  output: unknown;
  /** Structured error detail if the execution didn't succeed. Null otherwise. */
  error: unknown;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  finishedAt: string | null;
  /** @nullable */
  durationMs: number | null;
  retryCount: number;
  /**
     * Reserved for sub-workflow calls. Always null for now.
     * @nullable
     */
  parentId: string | null;
  createdAt: string;
}

export type ExecutionLogStatus = typeof ExecutionLogStatus[keyof typeof ExecutionLogStatus];


export const ExecutionLogStatus = {
  pending: 'pending',
  running: 'running',
  success: 'success',
  error: 'error',
  skipped: 'skipped',
} as const;

/**
 * A single node's run within an execution.
 */
export interface ExecutionLog {
  id: string;
  executionId: string;
  /** The graph node's `key` this log entry belongs to. */
  nodeKey: string;
  status: ExecutionLogStatus;
  /** The value passed into this node. Arbitrary JSON, or null. */
  input: unknown;
  /** The value this node produced. Null unless status is "success". */
  output: unknown;
  /** Structured error detail if this node failed. Null otherwise. */
  error: unknown;
  /**
     * Reserved for offloading large (>100 KB) payloads to object storage. Always null for now.
     * @nullable
     */
  storageRef: string | null;
  /** @nullable */
  durationMs: number | null;
  /** Always 1 — retries are not implemented yet. */
  attempt: number;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  finishedAt: string | null;
  createdAt: string;
}

export type ListWorkflowsParams = {
/**
 * Case-insensitive substring match on workflow name.
 */
search?: string;
/**
 * Filter by active status.
 */
isActive?: boolean;
/**
 * Only return workflows that have ALL of the given tags.
 */
tags?: string[];
/**
 * Opaque pagination cursor from a previous response's nextCursor.
 */
after?: string;
/**
 * Max results per page (1-100).
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListWorkflows200 = {
  workflows: Workflow[];
  /** @nullable */
  nextCursor: string | null;
  total: number;
};

export type CreateWorkflow201 = {
  workflow: Workflow;
  version: WorkflowVersion;
};

export type GetWorkflow200 = {
  workflow: Workflow;
  activeVersion: WorkflowVersion | null;
};

export type SaveWorkflowVersion200 = {
  workflow: Workflow;
  version: WorkflowVersionSummary;
};

export type UpdateWorkflow200 = {
  workflow: Workflow;
};

export type ListWorkflowVersions200 = {
  versions: WorkflowVersionSummary[];
};

export type RestoreWorkflowVersion200 = {
  workflow: Workflow;
};

export type ExecuteWorkflowBody = {
  /** Arbitrary JSON passed as the input to the workflow's entry node. */
  triggerPayload?: unknown;
};

export type ExecuteWorkflow202 = {
  execution: Execution;
};

export type ListExecutionsParams = {
/**
 * Only return executions for this workflow.
 */
workflowId?: string;
/**
 * Filter by execution status.
 */
status?: ExecutionStatus;
/**
 * Opaque pagination cursor from a previous response's nextCursor.
 */
after?: string;
/**
 * Max results per page (1-100).
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListExecutions200 = {
  executions: Execution[];
  /** @nullable */
  nextCursor: string | null;
  total: number;
};

export type GetExecution200 = {
  execution: Execution;
  logs: ExecutionLog[];
};

export type CancelExecution200 = {
  execution: Execution;
};