# lib/api-zod/src/generated/types/execution.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionStatus } from './executionStatus';
import type { ExecutionTriggerType } from './executionTriggerType';

/**
 * A single run of a workflow.
 */
export interface Execution {
  id: string;
  workflowId: string;
  versionId: string;
  status: ExecutionStatus;
  triggerType: ExecutionTriggerType;
  /** The input the entry node ran with. Arbitrary JSON, or null. */
  triggerPayload: unknown;
  /** The workflow's final output (its terminal node's output, or one merged object keyed by node if it has several). Null until the execution finishes. */
  output: unknown;
  /** Structured error detail if the execution didn't succeed. Null otherwise. */
  error: unknown;
  /** @nullable */
  startedAt: Date | null;
  /** @nullable */
  finishedAt: Date | null;
  /** @nullable */
  durationMs: number | null;
  retryCount: number;
  /**
     * Reserved for sub-workflow calls. Always null for now.
     * @nullable
     */
  parentId: string | null;
  createdAt: Date;
}