# lib/api-zod/src/generated/types/executionLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionLogStatus } from './executionLogStatus';

/**
 * A single node's run within an execution.
 */
export interface ExecutionLog {
  id: string;
  executionId: string;
  /** The graph node's `key` this log entry belongs to. */
  nodeKey: string;
  status: ExecutionLogStatus;
  /** The value passed into this node. Arbitrary JSON, or null. */
  input: unknown;
  /** The value this node produced. Null unless status is "success". */
  output: unknown;
  /** Structured error detail if this node failed. Null otherwise. */
  error: unknown;
  /**
     * Reserved for offloading large (>100 KB) payloads to object storage. Always null for now.
     * @nullable
     */
  storageRef: string | null;
  /** @nullable */
  durationMs: number | null;
  /** Always 1 — retries are not implemented yet. */
  attempt: number;
  /** @nullable */
  startedAt: Date | null;
  /** @nullable */
  finishedAt: Date | null;
  createdAt: Date;
}