# lib/api-zod/src/generated/types/executionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ExecutionStatus = typeof ExecutionStatus[keyof typeof ExecutionStatus];


export const ExecutionStatus = {
  pending: 'pending',
  running: 'running',
  success: 'success',
  error: 'error',
  cancelled: 'cancelled',
  timeout: 'timeout',
} as const;